feat: use new ModuleLoader APIs for init and unload
ModuleLoader now allows us to check if the module is
really pending unload in only then proceed to action.

Because these methods are public, they can be used by
any other module. This module is a playground for testing
new APIs and also an example of how to write really good
modules and flex with the APIs.